feat: Add initial date parameter to CalendarDatePickerState
This commit introduces a new optional parameter to the `CalendarDatePickerState` constructor. Users can now specify an initial `instant` (of type `kotlinx.datetime.Instant`) and `timeZone` (of type `kotlinx.datetime.TimeZone`) to set the starting date of the calendar.

If no values are provided, the calendar will default to the current system time and time zone, maintaining the previous behavior.